marche pas

+++ TP4/include/quicksort.h
@@ -7,7 +7,7 @@ typedef struct
     int limit_time;
 } process_t;
 
-typedef int (*value_function)(process_t *);
+typedef double (*value_function)(process_t *);
 void swap(process_t *a, process_t *b);
 int partition(process_t *arr, int low, int high, value_function func);
 void quicksort(process_t *arr, int low, int high, value_function func);
